:ambulance: feature(calendar): Update references and enhance styling in calendar.html
Removed the unused destination_repo setting in the sync_repo.yml to avoid redundancy. Styles are added to calendar.html for better color coding of events and the HTML has been structured for improved readability. Additionally, URLs and collections have been modified to be more readable and user friendly in the UI. These changes provide a better user experience.