:recycle: refactor(Setup): Add flourish-prn to requirements and update settings
This commit adds the 'flourish-prn' package to the list of requirements for the flourish-calendar app. This additional package extends the app's functionality, providing extra features that the app relies on. The settings file has been updated to provide more configuration options relevant to the app, such as 'DEVICE_ID' and 'SECRET_KEY.' Some applications were also reordered in the 'INSTALLED_APPS' list to ensure their correct loading during runtime. In .gitignore, a rule was added to ignore compiled files from the 'crypto_fields' library.