:ambulance: fix(child_medical_history_admin): Adjust child_medical_history_admin fields
Removed 'clinic_visit_child' field from admin model of child_medical_history_admin.py file in flourish_child application.

This commit is responsible for the removal of the field 'clinic_visit_child'. This change is made because it isn't being utilised in the admin model for the medical history of a child, therefore improving readability and maintainability of the code.