Moved 'export_events_as_csv' function from utils to views
The 'export_events_as_csv' function, along with its dependencies, was moved from 'utils/export_helper.py' to 'views.py'. We realized that as it tends to interact more with user requests and generate HTTP responses, having it within the views is more fitting to django's convention. The function's import statement has been accordingly updated in 'urls.py'. This refactor enhances the structure and readability of our code.