:ambulance: fix(Reminders): Update Reminder model to prevent duplicate reminders
This update prevents the creation of new duplicate reminders when reminder details are updated. It makes use of a field tracker in the Reminder model to trigger a function to remove duplicate reminders when certain fields are updated. Also, updates test to confirm the lack of duplications and checks reminder changes.